requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
cartopy>=0.22.0
google-generativeai>=0.3.0
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0 